+++ CHANGELOG.md
@@ -13,6 +13,7 @@ guarantees will be maintained within each major version series.
 ### Added
 - (Unindexed) FASTA reader & FastaSequence data structure.
 - Missing unit tests for internal BAM tag access.
+- Chemistry data for basecaller v3.3.
 
 ### Fixed
 - Reverse complement on padding base.